Algorithm moved into dedicated class
