+++ src/Intelliflo.SDK.Security/Algorithms/ISigningAlgorithm.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Intelliflo.SDK.Security.Algorithms
+{
+    internal interface ISigningAlgorithm
+    {
+        bool Verify(SignatureRequest arg);
+        Uri Sign(SignatureRequest arg);
+    }
+}
